+++ run.py
@@ -1,0 +1,1 @@
+from server.flaskr import app
