--- server/codes/Output/op_usn-1_1
@@ -1,2 +1,0 @@
-some 
-output here
