+++ server/flaskr/__init__.py
@@ -1,0 +1,9 @@
+from flask import Flask
+import sys
+import server.config as config
+
+app = Flask(__name__)
+app.config.from_object(config.DevelopmentConfig)
+
+from server.flaskr import apiServer 
+from server.flaskr import views
