+++ server/flaskr/apiClient.py
@@ -1,14 +1,7 @@
-#!../flask/bin/python3
-
 import requests as req
 import argparse
-# not working as expected :/
-# from titanium_silver import thread_custom 
 
 import json
-import sys
-sys.path.insert(0, '../titanium_silver')
-from thread_custom import threaded 
 
 def sendCode(inputCodeFilePath,questionHash):
 	# Open the code file using the path sent.
@@ -21,15 +14,14 @@ def sendCode(inputCodeFilePath,questionHash):
 	
 	# Create the data object to be sent to server.
 	data = {
-		"USN":"usn-1",
+		"USN":"someusn",
 		"code":code,
 		"progLang":"C++",
 		"questionHash":questionHash,
 	}
 
-	print(json.dumps(data))
 	# Make a post request with the data to be sent.
-	res = req.post("http://128.0.0.1:5000/submitCode",
+	res = req.post("http://127.0.0.1:5000/submitCode",
 			json = data,
 			headers={
 				"content-type":"application/json"
