+++ server/flaskr/test.cpp
@@ -1,0 +1,11 @@
+#include <iostream>
+
+int main(int args, char *argv[]) {
+	int i=1;
+	int j=1;
+	i++;
+	j++;
+	int res = i+j;
+	std::cout<<"Bye World:"<<res<<std::endl;
+}
+
